hw9

+++ HW9-executor/src/main/java/ru/otus/executor/Executor.java
@@ -2,11 +2,9 @@ package ru.otus.executor;
 
 import ru.otus.model.DataSet;
 
-import java.math.BigInteger;
-
 public interface Executor {
 
     <T extends DataSet> void save(T entity);
 
-    <T extends DataSet> T load(BigInteger id, Class<T> entityClass);
+    <T extends DataSet> T load(long id, Class<T> entityClass);
 }
